+++ CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 1.1.3 ] - 2024-06-29
+## [ 1.1.4 ] - 2024-06-29
+### Added
+- Copy constructor to ProcessingProgress.
+- Added pure string constructor to ProcessingUnitProgressFormatter.
 
 ## [ 1.1.2 ] - 2024-06-29
 ### Added
 - Added new interface for time measurement: IProcessingUnitRuntimeTimeMeasurement.
+- Added ProcessingUnitProgressFormatter.
 
 ### Changed
 - Added IProcessingUnitRuntimeTimeMeasurement to IProcessingUnitRunnableListener.